#53 moving ed25519 types in their own foler
