+++ cas-dotnet-sdk-tests/ED25519WrapperTests.cs
@@ -1,4 +1,5 @@
 ﻿using CasDotnetSdk.Signatures;
+using CasDotnetSdk.Signatures.Types;
 using System.Text;
 using Xunit;
 using static CasDotnetSdk.Signatures.ED25519Wrapper;
