+++ cas-dotnet-sdk/Signatures/ED25519Wrapper.cs
@@ -1,5 +1,6 @@
 ﻿using CasDotnetSdk.Http;
 using CasDotnetSdk.Signatures.Linux;
+using CasDotnetSdk.Signatures.Types;
 using CasDotnetSdk.Signatures.Windows;
 using CASHelpers;
 using CASHelpers.Types.HttpResponses.BenchmarkAPI;
@@ -18,38 +19,6 @@ namespace CasDotnetSdk.Signatures
         {
             this._platform = new OperatingSystemDeterminator().GetOperatingSystem();
             this._benchmarkSender = new BenchmarkSender();
-        }
-
-        public class Ed25519SignatureResult
-        {
-            public string Signature { get; set; }
-            public string PublicKey { get; set; }
-        }
-
-        public class Ed25519ByteSignatureResult
-        {
-            public byte[] Signature { get; set; }
-            public byte[] PublicKey { get; set; }
-        }
-
-        internal struct Ed25519SignatureStruct
-        {
-            public IntPtr Signature { get; set; }
-            public IntPtr Public_Key { get; set; }
-        }
-
-        internal struct Ed25519KeyPairBytesResultStruct
-        {
-            public IntPtr key_pair { get; set; }
-            public int length { get; set; }
-        }
-
-        internal struct Ed25519ByteSignatureResultStruct
-        {
-            public IntPtr signature_byte_ptr { get; set; }
-            public int signature_length { get; set; }
-            public IntPtr public_key { get; set; }
-            public int public_key_length { get; set; }
         }
 
         public byte[] GetKeyPairBytes()
