+++ cas-dotnet-sdk/Signatures/Types/Ed25519ByteSignatureResult.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CasDotnetSdk.Signatures.Types
+{
+    public class Ed25519ByteSignatureResult
+    {
+        public byte[] Signature { get; set; }
+        public byte[] PublicKey { get; set; }
+    }
+}
