+++ cas-dotnet-sdk/Signatures/Types/Ed25519SignatureResult.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CasDotnetSdk.Signatures.Types
+{
+    public class Ed25519SignatureResult
+    {
+        public string Signature { get; set; }
+        public string PublicKey { get; set; }
+    }
+}
