+++ cas-dotnet-sdk/Signatures/Windows/ED25519WindowsWrapper.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using CasDotnetSdk.Signatures.Types;
+using System;
 using System.Runtime.InteropServices;
 using static CasDotnetSdk.Signatures.ED25519Wrapper;
