+++ cas-dotnet-sdk/cas-dotnet-sdk.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
+		<TargetFrameworks>net8.0</TargetFrameworks>
 		<RootNamespace>CasDotnetSdk</RootNamespace>
 		<Nullable>enable</Nullable>
 		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
@@ -22,7 +22,6 @@
 
 	<ItemGroup>
 		<Folder Include="Hashers\" />
-		<Folder Include="Signatures\" />
 		<Folder Include="Symmetric\" />
 		<Folder Include="Storage\" />
 	</ItemGroup>
